requirements.txt
openai-whisper==20230918
requests~=2.31.0
flask~=3.0.0